cache key hash
